Core/PacketIO: Updated to 12.1.0
Port From (https://github.com/TrinityCore/TrinityCore/commit/e6dd6d8b96a9398d5f647e55883d260eef07213f)

+++ Source/Game/Networking/Packets/ChatPackets.cs
@@ -251,7 +251,7 @@ namespace Game.Networking.Packets
             _worldPacket.WriteUInt32(TargetVirtualAddress);
             _worldPacket.WriteUInt32(SenderVirtualAddress);
             _worldPacket.WriteUInt32(AchievementID);
-            _worldPacket.WriteUInt16((ushort)_ChatFlags);
+            _worldPacket.WriteUInt32((uint)_ChatFlags);
             _worldPacket.WriteFloat(DisplayTime);
             _worldPacket.WriteUInt32(SpellID);
             _worldPacket.WriteBits(SenderName.GetByteCount(), 11);
